requirements.txt
Flask
xmpppy
